Formally address DBAPI compiler warnings.
* Address warnings about type mismatches (mostly in 64-bit configurations)
  by harmonizing type usage and adding explicit casts where safe to do so.
* Retire unused member variables and methods.
* pythonpp_config.hpp: Avoid conflicts over HAVE_FSTAT, which ncbiconf*.h
  will start conditionally defining itself.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@70295 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/cache/test/dbapi_cache_test.cpp
@@ -139,8 +139,8 @@ int CDBAPI_CacheTest::Run(void)
                                                1,
                                                "");
 
-        int s = strlen(szTest);
-        assert(blob_size == size_t(s));
+        size_t s = strlen(szTest);
+        assert(blob_size == s);
 
         char blob_buf[1024] = {0,};
         blob_cache->Read("key_1",
@@ -155,7 +155,7 @@ int CDBAPI_CacheTest::Run(void)
         {
         auto_ptr<IWriter> wrt(blob_cache->GetWriteStream("key_3", 1, "sk1"));
         size_t bytes_written;
-        int s = strlen(szTest);
+        size_t s = strlen(szTest);
         wrt->Write(szTest, s, &bytes_written);
         wrt->Flush();
         }
